requirements.txt
discord.py
mongoengine
python-dotenv
pika
